refactor: extract shared mini event-bus mock into __mocks__/mini-event-bus.ts
Address PR review comment: the mini event-bus pattern (Map-based
emit→on delivery) was duplicated across four test harnesses.
Extracted into a shared createMiniEventBus() helper under
src/extensions/__mocks__/ and replaced all inline implementations.
