+++ src/extensions/ferment/index.test.ts
@@ -8,6 +8,7 @@ import { FermentEventStore } from "../../ferment/event-store.js"
 import { clearFermentCache, FermentStorage } from "../../ferment/store.js"
 import type { Ferment } from "../../ferment/types.js"
 import { createContext } from "../__mocks__/context.js"
+import { createMiniEventBus } from "../__mocks__/mini-event-bus.js"
 import { globalTipRegistry } from "../tips/registry.js"
 import fermentExtension from "./index.js"
 import { clearAllLifecycleGuards } from "./lifecycle-obligation-guard.js"
@@ -63,10 +64,7 @@ function registerFermentExtension(runtime?: FermentRuntime, flagValues: Record<s
 	const commands = new Map<string, CommandHandler>()
 	const shortcuts = new Map<string, { description?: string; handler: ShortcutHandler }>()
 	const registeredFlags = new Set<string>()
-	// Real mini event-bus: the plan-review request/decision flow routes through
-	// events.emit → events.on channels; a vi.fn() stub would swallow every
-	// delivery, leaving TUI outcomes unable to reach the decision handler.
-	const eventHandlers = new Map<string, ((data: unknown) => unknown)[]>()
+	const { events } = createMiniEventBus()
 
 	const pi = {
 		on: (event: string, handler: EventHandler) => {
@@ -96,20 +94,7 @@ function registerFermentExtension(runtime?: FermentRuntime, flagValues: Record<s
 		appendEntry: vi.fn(),
 		sendMessage: vi.fn(),
 		sendUserMessage: vi.fn(),
-		events: {
-			emit: vi.fn((channel: string, data: unknown) => {
-				for (const handler of eventHandlers.get(channel) ?? []) handler(data)
-			}),
-			on: vi.fn((channel: string, handler: (data: unknown) => unknown) => {
-				const list = eventHandlers.get(channel) ?? []
-				list.push(handler)
-				eventHandlers.set(channel, list)
-				return () => {
-					const idx = list.indexOf(handler)
-					if (idx !== -1) list.splice(idx, 1)
-				}
-			}),
-		},
+		events,
 	} as unknown as ExtensionAPI
 
 	fermentExtension(pi, runtime)
