+++ src/extensions/ferment/scoping-flow.integration.test.ts
@@ -16,6 +16,7 @@ import { afterEach, beforeEach, describe, expect, it, vi } from "vitest"
 import { FermentEventStore } from "../../ferment/event-store.js"
 import { clearFermentCache } from "../../ferment/store.js"
 import { createContext } from "../__mocks__/context.js"
+import { createMiniEventBus } from "../__mocks__/mini-event-bus.js"
 import { createDefaultFermentRuntime, type FermentRuntime } from "./runtime.js"
 import { clearAllPendingScopes, getPendingScope, runScopingFlow } from "./scoping.js"
 import { clearAllScopingGates, clearAllStepStarts, setActive } from "./state.js"
@@ -40,9 +41,7 @@ function createHarness() {
 	const runtime: FermentRuntime = { ...createDefaultFermentRuntime(), getStorage: () => eventStorage }
 	const tools = new Map<string, RegisteredTool>()
 	let activeTools: string[] = []
-	// Real mini event-bus — propose_ferment_scoping emits plan-review requests
-	// over pi.events; without this the emit throws on a missing `.emit`.
-	const eventHandlers = new Map<string, ((data: unknown) => unknown)[]>()
+	const { events } = createMiniEventBus()
 
 	const pi = {
 		on: vi.fn(),
@@ -52,20 +51,7 @@ function createHarness() {
 		sendMessage: vi.fn(),
 		sendUserMessage: vi.fn(),
 		appendEntry: vi.fn(),
-		events: {
-			emit: vi.fn((channel: string, data: unknown) => {
-				for (const handler of eventHandlers.get(channel) ?? []) handler(data)
-			}),
-			on: vi.fn((channel: string, handler: (data: unknown) => unknown) => {
-				const list = eventHandlers.get(channel) ?? []
-				list.push(handler)
-				eventHandlers.set(channel, list)
-				return () => {
-					const idx = list.indexOf(handler)
-					if (idx !== -1) list.splice(idx, 1)
-				}
-			}),
-		},
+		events,
 		getActiveTools: vi.fn(() => activeTools),
 		getAllTools: vi.fn(() => Array.from(tools.keys()).map((name) => ({ name }))),
 		setActiveTools: vi.fn((names: string[]) => {
